262: fixed tests recieving args/flags from test harness
The ConfigView.run() and Logout.run() commands were being processed with the args from process.argv, which were the mocha args which are obviously not valid for the architect cli. Changes in this branch forced that validation to happen which in turned caused these two tests to begin to fail. This commit refactors the tests to use the fancy-tests for oclif, which isolates the process args from the test call itself.