Add edge type for Outputs and add them to the graph
This technically doesn't fully work because calling `addEdge` for an output edge targetting a service that already has a service edge will result in the output edge not being added. This is because the graph only supports a single edge between services, which is fine for representing interface dependencies, but needs to be expanded to support multiple output edges between a pair of services.